refactor analytics, fastidiously avoid uploading bad gzip files
